+++ supabase/migrations/20260604021341_added_permission_level.sql
@@ -1,0 +1,3 @@
+alter table "public"."users_projects" add column "permission_level" text not null;
+
+